# API.yml
openapi: 3.0.3
info:
  title: Sample API
  description: A sample API to demonstrate OpenAPI specification.
  version: 1.0.0
servers:
  - url: http://localhost:7071/
    description: Development server
  - url: https://api.example.com/v1
    description: Production server
...
paths:
...
  /permission:
    get:
      summary: Get all Permission
      tags:
        - Permission
      responses:
        '200':
          description: A list of Permission
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PermissionResponse'
    post:
      summary: Create a new permission
      tags:
        - Permission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartmentRequest'
      responses:
        '201':
          description: Department created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartmentResponse'
        '400':
          description: Invalid input'
  /permission/{userId}:
    post:
      summary: add  permission
      tags:
        - Permission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/DepartmentRequest'
      responses:
        '201':
          description: Permission added successfully
          content:
            application/json:
              schema:
                type: array
                items:
                 $ref: '#/components/schemas/PermissionResponse'
        '400':
          description: Invalid input'
    put:
      summary: Update permission
      tags:
        - Permission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/PermissionRequest'
      responses:
        '201':
          description: Permission updated successfully
          content:
            application/json:
              schema:
                type: array
                items:
                 $ref: '#/components/schemas/PermissionResponse'
        '400':
          description: Invalid input'
components:
  schemas:
...
    DepartmentRequest:
      type: object
      required:
        - name
        
      properties:
        name:
          type: string
          example: Jane Smith
    DepartmentResponse:
      type: object
      required:
        - id
        - name
        
      properties:

        departmentId:
          type: integer
          example: 1
        departmentName:
          type: string
          example: Jane Smith
    PermissionRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: read
    PermissionResponse:
      type: object
      required:
        - id
        - name
        
      properties:

        permissionId:
          type: integer
          example: 1
        permissionName:
          type: string
          example: read