# Employee_Management/src/main/java/com/mycompany/employee_management/permission/PermissionRequest.java
/*
 * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
 * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
 */
package com.mycompany.employee_management.permission;

import com.mycompany.employee_management.department.*;

/**
 *
 * @author sachi
 */
public class PermissionRequest {
    
    
    private String name;

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public PermissionRequest(String name) {
        this.name = name;
    }

    public PermissionRequest() {
    }
    
    
    
    
    
}